packages/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
bcrypt
pyjwt
sqlalchemy
psycopg2-binary
httpx
aiohttp
